my-react-app/backend/requirements.txt
absl-py==2.1.0
ale-py==0.10.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==1.6.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
fastapi==0.115.6
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
future==1.0.0
gast==0.6.0
google-api-core==2.19.2
google-api-python-client==2.123.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.67.1
gym==0.17.3
gym-notices==0.0.8
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httplib2==0.22.0
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras>=2.12.0
keras-rl2==1.0.5
keras-tuner==1.4.7
kiwisolver==1.4.7
kt-legacy==1.0.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.6.1
pyglet==1.5.0
Pygments==2.18.0
pyparsing==3.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
rl==3.2
rl-agents==0.1.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tensorboard>=2.12.0
tensorboard-data-server>=0.7.0
tensorflow>=2.12.0
tensorflow-io-gcs-filesystem>=0.30.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.16.0
xgboost==2.1.2
yfinance==0.2.51
